+++ src/math/softmax.ts
@@ -20,7 +20,7 @@ import * as util from '../util';
 
 import * as axis_util from './axis_util';
 import {doc, operation} from './decorators';
-import {Tensor, Scalar} from './tensor';
+import {Scalar, Tensor} from './tensor';
 
 export class Ops {
   /**
